requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
attrs==23.1.0
bitsandbytes==0.41.1
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
courlan==0.9.5
cramjam==2.7.0
cycler==0.12.1
cymem==2.0.8
DataProperty==1.0.1
datasets==2.15.0
dateparser==1.2.0
dill==0.3.7
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exllamav2==0.0.11
fastapi==0.105.0
fastparquet==2023.10.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.60.0
h11==0.14.0
hqq==0.1.1.post1
htmldate==1.6.1
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jusText==3.0.0
kiwisolver==1.4.5
langcodes==3.3.0
llama_cpp_python @ https://github.com/oobabooga/llama-cpp-python-cuBLAS-wheels/releases/download/metal/llama_cpp_python-0.2.25-cp311-cp311-macosx_14_0_arm64.whl#sha256=9992f1ef176227f22ce47db84275d165ed94abb20ec8c4d05d7e903b4cc0fea8
lm-eval==0.3.0
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mbstrdecoder==1.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.8.8
numpy==1.24.4
oauthlib==3.2.2
openai==1.6.1
optimum==1.16.1
orjson==3.9.10
packaging==23.2
pandas==2.1.4
pathvalidate==3.2.0
peft==0.7.1
Pillow==10.1.0
portalocker==2.8.2
preshed==3.0.9
protobuf==4.23.4
psutil==5.9.6
pyarrow==14.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycountry==23.12.11
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pytablewriter==1.2.0
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rouge-score==0.1.2
rpds-py==0.13.2
rsa==4.9
sacrebleu==1.5.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.39.0
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
srsly==2.4.8
starlette==0.27.0
sympy==1.12
tabledata==1.3.3
tcolorpy==0.1.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
thinc==8.2.2
threadpoolctl==3.2.0
timm==0.9.12
tld==0.13
tokenizers==0.15.0
toolz==0.12.0
torch==2.1.1
torch-grammar @ git+https://github.com/oobabooga/torch-grammar.git@82850b5383a629f3b0fa1fba7d8f2aba3185ddb2
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
tqdm-multiprocess==0.0.11
trafilatura==1.6.4
transformers==4.36.2
typepy==1.3.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
wandb==0.16.1
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
